requirements.txt
easydict==1.9
einops==0.4.1
flax==0.5.2
h5py==3.7.0
imageio==2.19.5
jax==0.3.15
matplotlib==3.5.1
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.2.0
PyYAML==6.0
requests==2.27.1
scikit_learn==1.1.1
scipy==1.8.0
thop==0.1.1.post2207130030
torch==1.11.0
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.12.0
tqdm==4.64.0
visdom==0.1.8.9
yacs==0.1.8
